Crate paths

Thin wrappers around std::path/camino::path, distinguishing between absolute and relative paths.

Structs

AbsPath

Wrapper around an absolute Utf8Path.

AbsPathBuf

Wrapper around an absolute Utf8PathBuf.

FromPathBufError

A possible error value while converting a PathBuf to a Utf8PathBuf.

FromPathError

A possible error value while converting a Path to a Utf8Path.

Iter

An iterator over the Utf8Components of a Utf8Path, as str slices.

ReadDirUtf8

Iterator over the entries in a directory.

RelPath

Wrapper around a relative Utf8Path.

RelPathBuf

Wrapper around a relative Utf8PathBuf.

Utf8Ancestors

An iterator over Utf8Path and its ancestors.

Utf8Components

An iterator over the Utf8Components of a Utf8Path.

Utf8DirEntry

Entries returned by the ReadDirUtf8 iterator.

Utf8Path

A slice of a UTF-8 path (akin to str).

Utf8PathBuf

An owned, mutable UTF-8 path (akin to String).

Utf8PrefixComponent

A structure wrapping a Windows path prefix as well as its unparsed string representation.

Enums

Utf8Component

A single component of a path.

Utf8Prefix

Windows path prefixes, e.g., C: or \\server\share.