paths

Enum Utf8Component

pub enum Utf8Component<'a> {
    Prefix(Utf8PrefixComponent<'a>),
    RootDir,
    CurDir,
    ParentDir,
    Normal(&'a str),
}

A single component of a path.

A Utf8Component roughly corresponds to a substring between path separators (/ or \).

This enum is created by iterating over Utf8Components, which in turn is created by the components method on Utf8Path.

Examples

use camino::{Utf8Component, Utf8Path};

let path = Utf8Path::new("/tmp/foo/bar.txt");
let components = path.components().collect::<Vec<_>>();
assert_eq!(&components, &[
    Utf8Component::RootDir,
    Utf8Component::Normal("tmp"),
    Utf8Component::Normal("foo"),
    Utf8Component::Normal("bar.txt"),
]);

Variants

Prefix(Utf8PrefixComponent<'a>)

A Windows path prefix, e.g., C: or \\server\share.

There is a large variety of prefix types, see Utf8Prefixโs documentation for more.

Does not occur on Unix.

RootDir

The root directory component, appears after any prefix and before anything else.

It represents a separator that designates that a path starts from root.

CurDir

A reference to the current directory, i.e., ..

ParentDir

A reference to the parent directory, i.e., ...

Normal(&'a str)

A normal component, e.g., a and b in a/b.

This variant is the most common one, it represents references to files or directories.

Implementations

impl<'a> Utf8Component<'a>

pub fn as_str(&self) -> &'a str

Extracts the underlying str slice.

Examples

use camino::Utf8Path;

let path = Utf8Path::new("./tmp/foo/bar.txt");
let components: Vec<_> = path.components().map(|comp| comp.as_str()).collect();
assert_eq!(&components, &[".", "tmp", "foo", "bar.txt"]);

pub fn as_os_str(&self) -> &'a OsStr

Extracts the underlying OsStr slice.

Examples

use camino::Utf8Path;

let path = Utf8Path::new("./tmp/foo/bar.txt");
let components: Vec<_> = path.components().map(|comp| comp.as_os_str()).collect();
assert_eq!(&components, &[".", "tmp", "foo", "bar.txt"]);