paths

Enum Utf8Prefix

pub enum Utf8Prefix<'a> {
    Verbatim(&'a str),
    VerbatimUNC(&'a str, &'a str),
    VerbatimDisk(u8),
    DeviceNS(&'a str),
    UNC(&'a str, &'a str),
    Disk(u8),
}

Windows path prefixes, e.g., C: or \\server\share.

Windows uses a variety of path prefix styles, including references to drive volumes (like C:), network shared folders (like \\server\share), and others. In addition, some path prefixes are โverbatimโ (i.e., prefixed with \\?\), in which case / is not treated as a separator and essentially no normalization is performed.

Examples

use camino::{Utf8Component, Utf8Path, Utf8Prefix};
use camino::Utf8Prefix::*;

fn get_path_prefix(s: &str) -> Utf8Prefix {
    let path = Utf8Path::new(s);
    match path.components().next().unwrap() {
        Utf8Component::Prefix(prefix_component) => prefix_component.kind(),
        _ => panic!(),
    }
}

assert_eq!(Verbatim("pictures"), get_path_prefix(r"\\?\pictures\kittens"));
assert_eq!(VerbatimUNC("server", "share"), get_path_prefix(r"\\?\UNC\server\share"));
assert_eq!(VerbatimDisk(b'C'), get_path_prefix(r"\\?\C:\"));
assert_eq!(DeviceNS("BrainInterface"), get_path_prefix(r"\\.\BrainInterface"));
assert_eq!(UNC("server", "share"), get_path_prefix(r"\\server\share"));
assert_eq!(Disk(b'C'), get_path_prefix(r"C:\Users\Rust\Pictures\Ferris"));

Variants

Verbatim(&'a str)

Verbatim prefix, e.g., \\?\cat_pics.

Verbatim prefixes consist of \\?\ immediately followed by the given component.

VerbatimUNC(&'a str, &'a str)

Verbatim prefix using Windowsโ Uniform Naming Convention, e.g., \\?\UNC\server\share.

Verbatim UNC prefixes consist of \\?\UNC\ immediately followed by the serverโs hostname and a share name.

VerbatimDisk(u8)

Verbatim disk prefix, e.g., \\?\C:.

Verbatim disk prefixes consist of \\?\ immediately followed by the drive letter and :.

DeviceNS(&'a str)

Device namespace prefix, e.g., \\.\COM42.

Device namespace prefixes consist of \\.\ immediately followed by the device name.

UNC(&'a str, &'a str)

Prefix using Windowsโ Uniform Naming Convention, e.g. \\server\share.

UNC prefixes consist of the serverโs hostname and a share name.

Disk(u8)

Prefix C: for the given disk drive.

Implementations

impl Utf8Prefix<'_>

pub fn is_verbatim(&self) -> bool

Determines if the prefix is verbatim, i.e., begins with \\?\.

Examples

use camino::Utf8Prefix::*;

assert!(Verbatim("pictures").is_verbatim());
assert!(VerbatimUNC("server", "share").is_verbatim());
assert!(VerbatimDisk(b'C').is_verbatim());
assert!(!DeviceNS("BrainInterface").is_verbatim());
assert!(!UNC("server", "share").is_verbatim());
assert!(!Disk(b'C').is_verbatim());