paths

Struct AbsPathBuf

pub struct AbsPathBuf(/* private fields */);

A Utf8PathBuf that is guaranteed to be absolute.

Implementations

impl AbsPathBuf

pub fn assert(path: Utf8PathBuf) -> AbsPathBuf

Wrap the given absolute path in AbsPathBuf

Panics

Panics if path is not absolute.

pub fn assert_utf8(path: PathBuf) -> AbsPathBuf

Wrap the given absolute path in AbsPathBuf

Panics

Panics if path is not absolute.

pub fn as_path(&self) -> &AbsPath

Coerces to an AbsPath slice.

Equivalent of Utf8PathBuf::as_path for AbsPathBuf.

pub fn pop(&mut self) -> bool

Equivalent of Utf8PathBuf::pop for AbsPathBuf.

Note that this wonโt remove the root component, so self will still be absolute.

pub fn push<P: AsRef<Utf8Path>>(&mut self, suffix: P)

Equivalent of PathBuf::push for AbsPathBuf.

Extends self with path.

If path is absolute, it replaces the current path.

On Windows:

if path has a root but no prefix (e.g., \windows), it replaces everything except for the prefix (if any) of self.

if path has a prefix but no root, it replaces self.

if self has a verbatim prefix (e.g. \\?\C:\windows) and path is not empty, the new path is normalized: all references to . and .. are removed.

pub fn join(&self, path: impl AsRef<Utf8Path>) -> Self