paths
Struct Utf8PathBuf

pub struct Utf8PathBuf(/* private fields */);

An owned, mutable UTF-8 path (akin to String).

This type provides methods like push and set_extension that mutate the path in place. It also implements Deref to Utf8Path, meaning that all methods on Utf8Path slices are available on Utf8PathBuf values as well.

Examples

You can use push to build up a Utf8PathBuf from components:

use camino::Utf8PathBuf;

let mut path = Utf8PathBuf::new();

path.push(r"C:\");
path.push("windows");
path.push("system32");

path.set_extension("dll");

However, push is best used for dynamic situations. This is a better way to do this when you know all of the components ahead of time:

use camino::Utf8PathBuf;

let path: Utf8PathBuf = [r"C:\", "windows", "system32.dll"].iter().collect();

We can still do better than this! Since these are all strings, we can use From::from:

use camino::Utf8PathBuf;

let path = Utf8PathBuf::from(r"C:\windows\system32.dll");

Which method works best depends on what kind of situation you’re in.

Implementations

impl Utf8PathBuf

pub fn new() -> Utf8PathBuf

Allocates an empty Utf8PathBuf.

Examples

use camino::Utf8PathBuf;

let path = Utf8PathBuf::new();

pub fn from_path_buf(path: PathBuf) -> Result<Utf8PathBuf, PathBuf>

Creates a new Utf8PathBuf from a PathBuf containing valid UTF-8 characters.

Errors with the original PathBuf if it is not valid UTF-8.

For a version that returns a type that implements std::error::Error, use the TryFrom<PathBuf> impl.

Examples

use camino::Utf8PathBuf;
use std::ffi::OsStr;
use std::os::unix::ffi::OsStrExt;
use std::path::PathBuf;

let unicode_path = PathBuf::from("/valid/unicode");
Utf8PathBuf::from_path_buf(unicode_path).expect("valid Unicode path succeeded");

// Paths on Unix can be non-UTF-8.
let non_unicode_str = OsStr::from_bytes(b"\xFF\xFF\xFF");
let non_unicode_path = PathBuf::from(non_unicode_str);
Utf8PathBuf::from_path_buf(non_unicode_path).expect_err("non-Unicode path failed");

pub fn into_std_path_buf(self) -> PathBuf

Converts a Utf8PathBuf to a PathBuf.

This is equivalent to the From<Utf8PathBuf> for PathBuf impl, but may aid in type inference.

Examples

use camino::Utf8PathBuf;
use std::path::PathBuf;

let utf8_path_buf = Utf8PathBuf::from("foo.txt");
let std_path_buf = utf8_path_buf.into_std_path_buf();
assert_eq!(std_path_buf.to_str(), Some("foo.txt"));

// Convert back to a Utf8PathBuf.
let new_utf8_path_buf = Utf8PathBuf::from_path_buf(std_path_buf).unwrap();
assert_eq!(new_utf8_path_buf, "foo.txt");

pub fn with_capacity(capacity: usize) -> Utf8PathBuf

Creates a new Utf8PathBuf with a given capacity used to create the internal PathBuf. See with_capacity defined on PathBuf.

Requires Rust 1.44 or newer.

Examples

use camino::Utf8PathBuf;

let mut path = Utf8PathBuf::with_capacity(10);
let capacity = path.capacity();

// This push is done without reallocating
path.push(r"C:\");

assert_eq!(capacity, path.capacity());

pub fn as_path(&self) -> &Utf8Path

Coerces to a Utf8Path slice.

Examples

use camino::{Utf8Path, Utf8PathBuf};

let p = Utf8PathBuf::from("/test");
assert_eq!(Utf8Path::new("/test"), p.as_path());

pub fn push(&mut self, path: impl AsRef<Utf8Path>)

Extends self with path.

If path is absolute, it replaces the current path.

On Windows:

if path has a root but no prefix (e.g., \windows), it replaces everything except for the prefix (if any) of self.

if path has a prefix but no root, it replaces self.

Examples

Pushing a relative path extends the existing path:

use camino::Utf8PathBuf;

let mut path = Utf8PathBuf::from("/tmp");
path.push("file.bk");
assert_eq!(path, Utf8PathBuf::from("/tmp/file.bk"));

Pushing an absolute path replaces the existing path:

use camino::Utf8PathBuf;

let mut path = Utf8PathBuf::from("/tmp");
path.push("/etc");
assert_eq!(path, Utf8PathBuf::from("/etc"));

pub fn pop(&mut self) -> bool

Truncates self to self.parent.

Returns false and does nothing if self.parent is None. Otherwise, returns true.

Examples

use camino::{Utf8Path, Utf8PathBuf};

let mut p = Utf8PathBuf::from("/spirited/away.rs");

p.pop();
assert_eq!(Utf8Path::new("/spirited"), p);
p.pop();
assert_eq!(Utf8Path::new("/"), p);

pub fn set_file_name(&mut self, file_name: impl AsRef<str>)

Updates self.file_name to file_name.

If self.file_name was None, this is equivalent to pushing file_name.

Otherwise it is equivalent to calling pop and then pushing file_name. The new path will be a sibling of the original path. (That is, it will have the same parent.)

Examples

use camino::Utf8PathBuf;

let mut buf = Utf8PathBuf::from("/");
assert_eq!(buf.file_name(), None);
buf.set_file_name("bar");
assert_eq!(buf, Utf8PathBuf::from("/bar"));
assert!(buf.file_name().is_some());
buf.set_file_name("baz.txt");
assert_eq!(buf, Utf8PathBuf::from("/baz.txt"));

pub fn set_extension(&mut self, extension: impl AsRef<str>) -> bool

Updates self.extension to extension.

Returns false and does nothing if self.file_name is None, returns true and updates the extension otherwise.

If self.extension is None, the extension is added; otherwise it is replaced.

Examples

use camino::{Utf8Path, Utf8PathBuf};

let mut p = Utf8PathBuf::from("/feel/the");

p.set_extension("force");
assert_eq!(Utf8Path::new("/feel/the.force"), p.as_path());

p.set_extension("dark_side");
assert_eq!(Utf8Path::new("/feel/the.dark_side"), p.as_path());

pub fn into_string(self) -> String

Consumes the Utf8PathBuf, yielding its internal String storage.

Examples

use camino::Utf8PathBuf;

let p = Utf8PathBuf::from("/the/head");
let s = p.into_string();
assert_eq!(s, "/the/head");

pub fn into_os_string(self) -> OsString

Consumes the Utf8PathBuf, yielding its internal OsString storage.

Examples

use camino::Utf8PathBuf;
use std::ffi::OsStr;

let p = Utf8PathBuf::from("/the/head");
let s = p.into_os_string();
assert_eq!(s, OsStr::new("/the/head"));

pub fn into_boxed_path(self) -> Box<Utf8Path>

Converts this Utf8PathBuf into a boxed Utf8Path.

pub fn capacity(&self) -> usize

Invokes capacity on the underlying instance of PathBuf.

Requires Rust 1.44 or newer.

pub fn clear(&mut self)

Invokes clear on the underlying instance of PathBuf.

Requires Rust 1.44 or newer.

pub fn reserve(&mut self, additional: usize)

Invokes reserve on the underlying instance of PathBuf.

Requires Rust 1.44 or newer.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

Invokes try_reserve on the underlying instance of PathBuf.

Requires Rust 1.63 or newer.

pub fn reserve_exact(&mut self, additional: usize)

Invokes reserve_exact on the underlying instance of PathBuf.

Requires Rust 1.44 or newer.

pub fn try_reserve_exact( &mut self, additional: usize, ) -> Result<(), TryReserveError>

Invokes try_reserve_exact on the underlying instance of PathBuf.

Requires Rust 1.63 or newer.

pub fn shrink_to_fit(&mut self)

Invokes shrink_to_fit on the underlying instance of PathBuf.

Requires Rust 1.44 or newer.

pub fn shrink_to(&mut self, min_capacity: usize)

Invokes shrink_to on the underlying instance of PathBuf.

Requires Rust 1.56 or newer.